Clarification for server.features missing 'hosts'
The implicit assumption in the code of ElectronX/ElectrumX and Fulcrum already is that a missing hosts key or a missing 'route' in `server.features` indicates the peer server no longer wishes to peer or that the route in question is not available for peering.  As such,  I added language to the `server.featutes` description to indicate this subtlety.