Added documentation for new method blockchain.address.get_scripthash
I went ahead and created this method because it might be useful for
clients to prefer to work with script hashes directly (since they are
unambiguous) whereas addresses have some ambiguity attached to them.

I documented this method and also fixed some typos and some nits in the
existing documentation.